Move operations into separate files
