+++ arches/management/commands/makedatamigrations.py
@@ -24,6 +24,10 @@ from django.apps import apps
 from django.conf import settings
 from django.core.management.base import BaseCommand, CommandError
 from arches.app.models.models import GraphXPublishedGraph, PublishedGraph
+from arches.db.migration_operations.add_node_to_tile_data import AddNodeToTileData
+from arches.db.migration_operations.create_graph import CreateGraph
+from arches.db.migration_operations.delete_node_from_tile_data import DeleteNodeFromTileData
+from arches.db.migration_operations.update_resource_instances_publication_id import UpdateResourceInstancesPublicationId
 
 
 class Command(BaseCommand):
@@ -488,12 +492,27 @@ class MigrationWriter:
         self.operations = operations
         self.dependencies = dependencies or []
 
+    # Maps operation class name → submodule within data_migration_operations package.
+    _CLASS_MODULE = {
+        "AddNodeToTileData": "add_node_to_tile_data",
+        "CreateGraph": "create_graph",
+        "DeleteNodeFromTileData": "delete_node_from_tile_data",
+        "UpdateGraphFromJSON": "update_graph_from_json",
+        "UpdateResourceInstancesPublicationId": "update_resource_instances_publication_id",
+    }
+
     def as_string(self) -> str:
         now = datetime.datetime.now(tz=datetime.timezone.utc).strftime(
             "%Y-%m-%d %H:%M:%S"
         )
         imports = self._collect_imports()
         ops_block = self._render_operations()
+
+        base_pkg = "arches.db.migration_operations"
+        import_lines = [
+            f"from {base_pkg}.{self._CLASS_MODULE[name]} import {name}"
+            for name in imports
+        ]
 
         lines = [
             f"# Generated by Arches makedatamigrations on {now}",
@@ -502,9 +521,7 @@ class MigrationWriter:
             f"# To publication:   {self.pub_b.publicationid}  [{self.pub_b.published_time}]",
             "",
             "from django.db import migrations",
-            "from arches.db.migration_operations.data_migration_operations import (",
-            *[f"    {name}," for name in imports],
-            ")",
+            *import_lines,
             "",
             "",
             "class Migration(migrations.Migration):",
@@ -547,46 +564,18 @@ class MigrationWriter:
         return "\n".join(rendered)
 
     def _render_op(self, op: dict) -> str:
+        pub_a_id = str(self.pub_a.publicationid)
         if op["op"] == "CreateGraph":
-            return "\n".join(
-                [
-                    "        CreateGraph(",
-                    f"            graph_id={op['graph_id']!r},",
-                    f"            graph_slug={op['graph_slug']!r},",
-                    f"            name={op['name']!r},",
-                    "        ),",
-                ]
-            )
+            return CreateGraph.as_migration_string(op)
         if op["op"] == "CreateNode" and op.get("nodegroup_is_existing"):
-            return "\n".join(
-                [
-                    "        AddNodeToTileData(",
-                    f"            publication_id={str(self.pub_a.publicationid)!r},",
-                    f"            nodegroup_id={op['nodegroup_id']!r},",
-                    f"            node_id={op['nodeid']!r},",
-                    f"            value={op.get('default_value')!r},",
-                    "        ),",
-                ]
-            )
+            return AddNodeToTileData.as_migration_string(op, pub_a_id)
         if op["op"] == "DeleteNode" and op.get("nodegroup_id"):
-            return "\n".join(
-                [
-                    "        DeleteNodeFromTileData(",
-                    f"            publication_id={str(self.pub_a.publicationid)!r},",
-                    f"            nodegroup_id={op['nodegroup_id']!r},",
-                    f"            node_id={op['nodeid']!r},",
-                    f"            value=None,  # TODO: set default value for {op['alias']!r} ({op['datatype']!r})",
-                    "        ),",
-                ]
-            )
+            return DeleteNodeFromTileData.as_migration_string(op, pub_a_id)
         return f"        # {_format_op_for_display(op)}"
 
     def _render_update_publication_op(self) -> str:
-        return "\n".join(
-            [
-                "        UpdateResourceInstancesPublicationId(",
-                f"            current_publication_id={str(self.pub_a.publicationid)!r},",
-                f"            updated_publication_id={str(self.pub_b.publicationid)!r},",
-                "        ),",
-            ]
+        return UpdateResourceInstancesPublicationId.as_migration_string(
+            str(self.pub_a.publicationid),
+            str(self.pub_b.publicationid),
         )
+
